Move all components to batch-flask activity module
