+++ app/app.routes.ts
@@ -1,7 +1,7 @@
 // tslint:disable: object-literal-sort-keys
 import { Routes } from "@angular/router";
 
-import { ActivityMonitorComponent } from "app/components/activity-monitor";
+import { ActivityMonitorComponent } from "@batch-flask/ui/activity/activity-monitor";
 import { NavigationGuard } from "app/components/common/guards";
 import { JobGraphsComponent } from "app/components/job/graphs/job-graphs-home";
 import { ChooseActionComponent } from "app/components/market/application-action";
