--- app/components/activity-monitor/index.ts
@@ -1,2 +1,0 @@
-export * from "./activity-monitor.module";
-export * from "./activity-monitor.component";
