+++ src/@batch-flask/ui/activity/activity-monitor-footer-item/activity-monitor-footer-item.html
