+++ src/@batch-flask/ui/activity/activity-monitor-footer/activity-monitor-footer.scss
