+++ src/@batch-flask/ui/activity/activity-monitor/activity-monitor-item/activity-monitor-item-action/activity-monitor-item-action.component.ts
@@ -8,6 +8,7 @@ export interface ActivityAction {
     action: () => void;
     faClass: string;
     active: boolean;
+    id: number;
 }
 
 @Component({
