+++ src/@batch-flask/ui/activity/activity-monitor/activity-monitor-item/activity-monitor-item-action/activity-monitor-item-action.scss
