+++ src/@batch-flask/ui/activity/activity-monitor/activity-monitor-item/activity-monitor-item.scss
