+++ src/@batch-flask/ui/activity/activity-monitor/activity-monitor-item/index.ts
@@ -1,1 +1,2 @@
 export * from "./activity-monitor-item.component";
+export * from "./activity-monitor-item-action";
