+++ src/@batch-flask/ui/activity/activity-monitor/activity-monitor.scss
