+++ src/@batch-flask/ui/activity/index.ts
@@ -1,6 +1,7 @@
-export * from "./activity";
+export * from "./activity-types";
 export * from "./activity.service";
 export * from "./activity-monitor-footer-item";
 export * from "./activity-monitor-footer";
 export * from "./activity-history-queue";
 export * from "./activity.module";
+export * from "./activity-monitor";
